Removed Duplication from TestBatch.java
