+++ bigtable-hbase-parent/bigtable-hbase-integration-tests-common/src/main/java/com/google/cloud/bigtable/hbase/AbstractTestBatch.java
@@ -23,7 +23,9 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
 
+import org.apache.commons.lang.ArrayUtils;
 import org.apache.hadoop.hbase.CellUtil;
+import org.apache.hadoop.hbase.client.Append;
 import org.apache.hadoop.hbase.client.Connection;
 import org.apache.hadoop.hbase.client.Delete;
 import org.apache.hadoop.hbase.client.Get;
@@ -340,4 +342,48 @@ public abstract class AbstractTestBatch extends AbstractTest {
     } catch(Exception e) {
     }
   }
+  
+  /**
+   * Requirement 8.1
+   */
+  @Test
+  public void testBatchAppend() throws IOException, InterruptedException {
+    // Initialize data
+    Table table = getDefaultTable();
+    byte[] rowKey1 = dataHelper.randomData("testrow-");
+    byte[] qual1 = dataHelper.randomData("qual-");
+    byte[] value1_1 = dataHelper.randomData("value-");
+    byte[] value1_2 = dataHelper.randomData("value-");
+    byte[] rowKey2 = dataHelper.randomData("testrow-");
+    byte[] qual2 = dataHelper.randomData("qual-");
+    byte[] value2_1 = dataHelper.randomData("value-");
+    byte[] value2_2 = dataHelper.randomData("value-");
+
+    // Put
+    Put put1 = new Put(rowKey1).addColumn(COLUMN_FAMILY, qual1, value1_1);
+    Put put2 = new Put(rowKey2).addColumn(COLUMN_FAMILY, qual2, value2_1);
+    List<Row> batch = new ArrayList<Row>(2);
+    batch.add(put1);
+    batch.add(put2);
+    table.batch(batch, null);
+
+    // Increment
+    Append append1 = new Append(rowKey1);
+    appendAdd(append1, COLUMN_FAMILY, qual1, value1_2);
+    Append append2 = new Append(rowKey2);
+    appendAdd(append2, COLUMN_FAMILY, qual2, value2_2);
+    batch.clear();
+    batch.add(append1);
+    batch.add(append2);
+    Object[] results = new Object[2];
+    table.batch(batch, results);
+    Assert.assertArrayEquals("Should be value1_1 + value1_2", ArrayUtils.addAll(value1_1, value1_2),
+      CellUtil.cloneValue(((Result) results[0]).getColumnLatestCell(COLUMN_FAMILY, qual1)));
+    Assert.assertArrayEquals("Should be value1_1 + value1_2", ArrayUtils.addAll(value2_1, value2_2),
+      CellUtil.cloneValue(((Result) results[1]).getColumnLatestCell(COLUMN_FAMILY, qual2)));
+
+    table.close();
+  }
+  
+  protected abstract void appendAdd(Append append, byte[] columnFamily, byte[] qualifier, byte[] value);
 }
